+++ kernel/src/HAL/hal.cpp
@@ -1,5 +1,3 @@
-#include "timer.hpp"
-
 #include <arch/x86_64/GDT/gdt.hpp>
 
 #include <arch/x86_64/interrupts/IDT.hpp>
@@ -14,11 +12,17 @@
 
 #include <arch/x86_64/Memory/PagingInit.hpp>
 
+#include "drivers/ACPI/RSDP.hpp"
+#include "drivers/ACPI/XSDT.hpp"
+
+#include "timer.hpp"
 #include "hal.hpp"
+
+#include <assert.h>
 
 namespace WorldOS {
 
-    void HAL_Init(MemoryMapEntry** MemoryMap, uint64_t MMEntryCount, uint64_t kernel_virtual, uint64_t kernel_physical, uint64_t kernel_size, uint64_t HHDM_start, const FrameBuffer& fb) {
+    void HAL_EarlyInit(MemoryMapEntry** MemoryMap, uint64_t MMEntryCount, uint64_t kernel_virtual, uint64_t kernel_physical, uint64_t kernel_size, uint64_t HHDM_start, const FrameBuffer& fb) {
         GDT* gdt = &DefaultGDT;
         GDTDescriptor gdtDescriptor = {(sizeof(GDT) - 1), ((uint64_t)gdt)};
         x86_64_LoadGDT(&gdtDescriptor);
@@ -41,6 +45,12 @@ namespace WorldOS {
         x86_64_InitPaging(MemoryMap, MMEntryCount, kernel_virtual, kernel_physical, kernel_size, (uint64_t)(fb.FrameBufferAddress), ((fb.bpp >> 3) * fb.FrameBufferHeight * fb.FrameBufferWidth), HHDM_start);
 
         x86_64_EnableInterrupts();
+    }
+
+    void HAL_Stage2(void* RSDP) {
+        assert(InitAndValidateRSDP(RSDP));
+        assert(IsXSDTAvailable());
+        assert(InitAndValidateXSDT(GetXSDT()));
     }
 
     /*